requirements.txt
# Temel bağımlılıklar
streamlit==1.44.0
langchain==0.3.23
langgraph==0.3.27
gspread==6.2.0
oauth2client==4.1.3
pandas==2.2.3
python-dotenv==1.0.1

# Google API bağımlılıkları
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# HTTP ve web bağımlılıkları
requests==2.32.3
aiohttp==3.11.14
httpx==0.28.1

# Veri işleme
numpy==1.26.4
pydantic==2.11.2
mcp